requirements.txt
streamlit>=1.42.0
Authlib>=1.3.2
bcrypt
pandas
python-dotenv
pydantic
pydantic[email]
streamlit-google-auth==1.1.8
sqlalchemy
psycopg2-binary